HTMLEntities.Decoder

Description

Utilities, which execute the parser.

Documentation

htmlEntity :: Text -> Either String Text

A decoder of a single entity.

>>> mapM_ Data.Text.IO.putStrLn $ htmlEntity "&#169;"
©

htmlEntityBody :: Text -> Either String Text

A decoder of a single entity.

>>> mapM_ Data.Text.IO.putStrLn $ htmlEntityBody "#169"
©

htmlEncodedText :: Text -> Builder

A decoder of a text with entities.

Produces a text builder, which you can then convert into a text or a lazy text, using the "text" or "conversion-text" library.

>>> Data.Text.Lazy.IO.putStrLn $ Data.Text.Lazy.Builder.toLazyText $ htmlEncodedText "&euro;5 &cent;2"
€5 ¢2